+++ skaff/clitools.py
@@ -4,6 +4,7 @@
 A suite of command line based tools.
 """
 # --------------------------------- MODULES -----------------------------------
+import argparse
 import errno
 import fcntl
 import os
@@ -15,6 +16,7 @@ from functools import wraps
 # --------------------------------- MODULES -----------------------------------
 
 
+# --------------------------------- CLASSES -----------------------------------
 class ANSIColor:
     RED = "\x1b[31m"
     GREEN = "\x1b[32m"
@@ -32,6 +34,70 @@ class TimeOutError(Exception):
     pass
 
 
+class SmartFormatter(argparse.HelpFormatter):
+    """
+    You can only specify one formatter in standard argparse, so you cannot
+    both have pre-formatted description (RawDescriptionHelpFormatter)
+    and ArgumentDefaultsHelpFormatter.
+    The SmartFormatter has sensible defaults (RawDescriptionFormatter) and
+    the individual help text can be marked ( help="R|" ) for
+    variations in formatting.
+    Version string is formatted using _split_lines and preserves any
+    line breaks in the version string.
+    """
+    # Use a custom formatter to ensure the custom formatted version string
+    # got printed properly while still preserves line-wrapping and other
+    # functionality for all the rest of the help text (argparse)
+    # Borrowed from
+    # https://bitbucket.org/ruamel/std.argparse/overview
+    def __init__(self, *args, **kwargs):
+        self._add_defaults = None
+        super(SmartFormatter, self).__init__(*args, **kwargs)
+
+    def _fill_text(self, text, width, indent):
+        return ''.join([indent + line for line in text.splitlines(True)])
+
+    def _split_lines(self, text, width):
+        if text.startswith('D|'):
+            self._add_defaults = True
+            text = text[2:]
+        elif text.startswith('*|'):
+            text = text[2:]
+        if text.startswith('R|'):
+            return text[2:].splitlines()
+        return argparse.HelpFormatter._split_lines(self, text, width)
+
+    def _get_help_string(self, action):
+        if self._add_defaults is None:
+            return argparse.HelpFormatter._get_help_string(self, action)
+        help = action.help
+        if '%(default)' not in action.help:
+            if action.default is not argparse.SUPPRESS:
+                defaulting_nargs = [argparse.OPTIONAL, argparse.ZERO_OR_MORE]
+                if action.option_strings or action.nargs in defaulting_nargs:
+                    help += ' (default: %(default)s)'
+        return help
+
+    def _expand_help(self, action):
+        """
+        Mark a password help with '*|' at the start, so that
+        when global default adding is activated (e.g. through a helpstring
+        starting with 'D|') no password is show by default.
+        Orginal marking used in repo cannot be used because of decorators.
+        """
+        hs = self._get_help_string(action)
+        if hs.startswith('*|'):
+            params = dict(vars(action), prog=self._prog)
+            if params.get('default') is not None:
+                # you can update params, this will change the default, but we
+                # are printing help only
+                params['default'] = '*' * len(params['default'])
+            return self._get_help_string(action) % params
+        return super(SmartFormatter, self)._expand_help(action)
+# --------------------------------- CLASSES -----------------------------------
+
+
+# -------------------------------- FUNCTIONS ----------------------------------
 def timeout(seconds, error_message=os.strerror(errno.ETIME)):
     """Sets a timer on a function; should be used as a decorator.
 
@@ -112,3 +178,4 @@ def getkey():
         termios.tcsetattr(fd, termios.TCSAFLUSH, attrs_save)
         fcntl.fcntl(fd, fcntl.F_SETFL, flags_save)
     return ret
+# -------------------------------- FUNCTIONS ----------------------------------